Refactor to use split up use cases
