+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/Main.kt
@@ -1,23 +1,17 @@
 package org.wycliffeassociates.otter.jvm.app
 
-import org.wycliffeassociates.otter.common.data.model.Collection
-import org.wycliffeassociates.otter.common.data.model.Language
-import org.wycliffeassociates.otter.common.data.model.ResourceMetadata
-import org.wycliffeassociates.otter.common.domain.ImportAudioPlugins
+import org.wycliffeassociates.otter.common.domain.plugins.ImportAudioPlugins
 import org.wycliffeassociates.otter.common.domain.ImportLanguages
-import org.wycliffeassociates.otter.common.domain.PluginActions
+import org.wycliffeassociates.otter.common.domain.plugins.InitializePlugins
 import org.wycliffeassociates.otter.jvm.app.ui.inject.Injector
 import org.wycliffeassociates.otter.jvm.app.ui.menu.MainMenu
 import org.wycliffeassociates.otter.jvm.app.ui.menu.MainMenuStylesheet
 import org.wycliffeassociates.otter.jvm.app.ui.projecthome.ProjectHomeView
-import org.wycliffeassociates.otter.jvm.app.ui.projectpage.view.ProjectPage
 import org.wycliffeassociates.otter.jvm.app.ui.projectpage.view.ProjectPageStylesheet
 import org.wycliffeassociates.otter.jvm.app.ui.styles.AppStyles
 import org.wycliffeassociates.otter.jvm.app.ui.viewtakes.view.ViewTakesStylesheet
-import org.wycliffeassociates.otter.jvm.persistence.DefaultPluginPreference
 import tornadofx.*
 import java.io.File
-import java.time.LocalDate
 
 class MyApp : App(Workspace::class) {
     init {
@@ -49,6 +43,6 @@ private fun initApp() {
 
     ImportAudioPlugins(Injector.audioPluginRegistrar, Injector.directoryProvider)
             .importAll()
-            .andThen(PluginActions(Injector.pluginRepository).initializeDefault())
+            .andThen(InitializePlugins(Injector.pluginRepository).initDefault())
             .subscribe()
 }
