+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/menu/MainMenu.kt
@@ -7,10 +7,9 @@ import io.reactivex.schedulers.Schedulers
 import javafx.scene.control.MenuBar
 import javafx.scene.control.ToggleGroup
 import org.wycliffeassociates.otter.common.domain.ImportResourceContainer
-import org.wycliffeassociates.otter.common.domain.PluginActions
+import org.wycliffeassociates.otter.common.domain.plugins.AccessPlugins
 import org.wycliffeassociates.otter.jvm.app.ui.inject.Injector
 import tornadofx.*
-
 
 class MainMenu : MenuBar() {
 
@@ -47,8 +46,8 @@ class MainMenu : MenuBar() {
                     val pluginToggleGroup = ToggleGroup()
 
                     // Get the plugins from the use case
-                    val pluginActions = PluginActions(pluginRepository)
-                    pluginActions
+                    val accessPlugins = AccessPlugins(pluginRepository)
+                    accessPlugins
                             .getAllPluginData()
                             .observeOnFx()
                             .doOnSuccess { pluginData ->
@@ -56,7 +55,7 @@ class MainMenu : MenuBar() {
                                     radiomenuitem(it.name) {
                                         userData = it
                                         action {
-                                            pluginActions.setDefaultPluginData(it).subscribe()
+                                            accessPlugins.setDefaultPluginData(it).subscribe()
                                         }
                                         toggleGroup = pluginToggleGroup
                                     }
@@ -64,7 +63,7 @@ class MainMenu : MenuBar() {
                             }
                             // Select the default plugin
                             .flatMapMaybe {
-                                pluginActions.getDefaultPluginData()
+                                accessPlugins.getDefaultPluginData()
                             }
                             .observeOnFx()
                             .subscribe { plugin ->
