+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/device/audioplugin/AudioPluginRegistrar.kt
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import io.reactivex.Completable
 import io.reactivex.Observable
 import io.reactivex.Single
-import org.wycliffeassociates.otter.common.domain.IAudioPluginRegistrar
+import org.wycliffeassociates.otter.common.domain.plugins.IAudioPluginRegistrar
 import org.wycliffeassociates.otter.common.persistence.repositories.IAudioPluginRepository
 import org.wycliffeassociates.otter.jvm.device.audioplugin.parser.ParsedAudioPluginDataMapper
 import java.io.File
