+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/device/audioplugin/injection/AudioPluginComponent.kt
@@ -1,7 +1,7 @@
 package org.wycliffeassociates.otter.jvm.device.audioplugin.injection
 
 import dagger.Component
-import org.wycliffeassociates.otter.common.domain.IAudioPluginRegistrar
+import org.wycliffeassociates.otter.common.domain.plugins.IAudioPluginRegistrar
 import javax.inject.Singleton
 
 @Component(modules = [AudioPluginModule::class])
